% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruise.R
\name{cruise}
\alias{cruise}
\title{Forest cruise input}
\usage{
cruise(path)
}
\arguments{
\item{path}{path to .xls or .xlsx file with inventory data. See example
workbook for proper formatting.}
}
\value{
cruise object
}
\description{
Process forest inventory data from a properly formatted Excel workbook.
}
