% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{due_every}
\alias{due_every}
\title{Due Every}
\format{A data frame with several items and mileage due for each item.
\describe{
  \item{Item}{Description of maintenanced item}
  \item{mileage_done}{Mileage at which the item was maintenanced}
}}
\description{
Dataset contains a list of maintenance items and recommended
mileage for doing the maintenance
}
