% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_max_mileage}
\alias{find_max_mileage}
\title{Find max milage}
\usage{
find_max_mileage(dtfr)
}
\arguments{
\item{drfr}{a data frame containing history for one maintenance item}
}
\value{
dataframe containing the item and last mileage maintenace was done
(mileage_done)
}
\description{
functions takes current maintenance history for a single item
and finds the last mileage maintenance was done (i.e. max mileage)
}
