% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maintenance_history}
\alias{maintenance_history}
\title{Maintenance History of my LX470}
\format{A data frame with several items and history of when the milage was
done
\describe{
  \item{Item}{Description of maintenanced item}
  \item{mileage_done}{Mileage at which the item was maintenanced}
}}
\description{
Dataset contains the maintenence items as well as mileage the item
was completed.
}
